+++ app/Werewolf/Commands/See.hs
@@ -46,6 +46,6 @@ handle callerName (Options targetName) = do
 
     let command = seeCommand callerName targetName
 
-    case runExcept (runWriterT $ execStateT (apply command >> checkTurn >> checkGameOver) game) of
+    case runExcept (runWriterT $ execStateT (apply command >> checkStage >> checkGameOver) game) of
         Left errorMessages      -> exitWith failure { messages = errorMessages }
         Right (game', messages) -> writeGame game' >> exitWith success { messages = messages }