+++ test/src/Game/Werewolf/Test/Engine.hs
@@ -9,8 +9,8 @@ Maintainer  : public@hjwylde.com
 {-# LANGUAGE OverloadedStrings #-}
 
 module Game.Werewolf.Test.Engine (
-    -- * checkTurn
-    prop_checkTurnSkipsSeersWhenNoSeers, prop_checkTurnDoesNothingWhenGameOver,
+    -- * checkStage
+    prop_checkStageSkipsSeersWhenNoSeers, prop_checkStageDoesNothingWhenGameOver,
     prop_checkSeersTurnAdvancesToWerewolves, prop_checkSeersTurnResetsSees,
     prop_checkSeersTurnDoesNothingUnlessAllSeen, prop_checkVillagersTurnAdvancesToSeers,
     prop_checkVillagersTurnLynchesOnePlayerWhenConsensus, prop_checkVillagersTurnLynchesNoOneWhenConflictedAndNoScapegoats,
@@ -21,10 +21,10 @@ module Game.Werewolf.Test.Engine (
     prop_checkWerewolvesTurnDoesNothingUnlessAllVoted,
 
     -- * checkGameOver
-    prop_checkGameOverAdvancesTurn, prop_checkGameOverDoesNothingWhenAtLeastTwoAllegiancesAlive,
+    prop_checkGameOverAdvancesStage, prop_checkGameOverDoesNothingWhenAtLeastTwoAllegiancesAlive,
 
     -- * startGame
-    prop_startGameStartsWithNightfallTurn, prop_startGameUsesGivenPlayers,
+    prop_startGameStartsWithSunsetStage, prop_startGameUsesGivenPlayers,
     prop_startGameErrorsUnlessUniquePlayerNames, prop_startGameErrorsWhenLessThan7Players,
     prop_startGameErrorsWhenMoreThan24Players,
 
@@ -57,68 +57,68 @@ import           Game.Werewolf.Test.Arbitrary
 import Test.QuickCheck
 import Test.QuickCheck.Monadic
 
-prop_checkTurnSkipsSeersWhenNoSeers :: Game -> Property
-prop_checkTurnSkipsSeersWhenNoSeers game =
+prop_checkStageSkipsSeersWhenNoSeers :: Game -> Property
+prop_checkStageSkipsSeersWhenNoSeers game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isWerewolvesTurn $ run_ checkTurn game''
+    isWerewolvesTurn $ run_ checkStage game''
     where
-        game'   = (foldl killPlayer game (filterSeers $ game ^. players)) { _turn = Villagers }
+        game'   = (foldl killPlayer game (filterSeers $ game ^. players)) { _stage = VillagersTurn }
         n       = length . filterAlive $ game' ^. players
 
-prop_checkTurnDoesNothingWhenGameOver :: Game -> Property
-prop_checkTurnDoesNothingWhenGameOver game = run_ checkTurn game' === game'
+prop_checkStageDoesNothingWhenGameOver :: Game -> Property
+prop_checkStageDoesNothingWhenGameOver game = run_ checkStage game' === game'
     where
-        game' = game { _turn = NoOne }
+        game' = game { _stage = GameOver }
 
 prop_checkSeersTurnAdvancesToWerewolves :: Game -> Property
 prop_checkSeersTurnAdvancesToWerewolves game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isWerewolvesTurn $ run_ checkTurn game''
+    isWerewolvesTurn $ run_ checkStage game''
     where
-        game'   = game { _turn = Seers }
+        game'   = game { _stage = SeersTurn }
         n       = length . filterSeers $ game' ^. players
 
 prop_checkSeersTurnResetsSees :: Game -> Property
 prop_checkSeersTurnResetsSees game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    Map.null $ run_ checkTurn game'' ^. sees
+    Map.null $ run_ checkStage game'' ^. sees
     where
-        game'   = game { _turn = Seers }
+        game'   = game { _stage = SeersTurn }
         n       = length . filterSeers $ game' ^. players
 
 prop_checkSeersTurnDoesNothingUnlessAllSeen :: Game -> Property
 prop_checkSeersTurnDoesNothingUnlessAllSeen game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isSeersTurn $ run_ checkTurn game''
+    isSeersTurn $ run_ checkStage game''
     where
-        game'   = game { _turn = Seers }
+        game'   = game { _stage = SeersTurn }
         n       = length (filterSeers $ game' ^. players) - 1
 
 prop_checkVillagersTurnAdvancesToSeers :: Game -> Property
 prop_checkVillagersTurnAdvancesToSeers game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    not (null . filterAlive . filterSeers $ run_ checkTurn game'' ^. players)
-    ==> isSeersTurn $ run_ checkTurn game''
+    not (null . filterAlive . filterSeers $ run_ checkStage game'' ^. players)
+    ==> isSeersTurn $ run_ checkStage game''
     where
-        game'   = game { _turn = Villagers }
+        game'   = game { _stage = VillagersTurn }
         n       = length $ game' ^. players
 
 prop_checkVillagersTurnLynchesOnePlayerWhenConsensus :: Game -> Property
 prop_checkVillagersTurnLynchesOnePlayerWhenConsensus game =
     forAll (runArbitraryCommands n game') $ \game'' ->
     length (last $ groupSortOn (length . flip elemIndices (Map.elems $ game'' ^. votes)) (nub . Map.elems $ game'' ^. votes)) == 1
-    ==> length (filterDead $ run_ checkTurn game'' ^. players) == 1
+    ==> length (filterDead $ run_ checkStage game'' ^. players) == 1
     where
-        game'   = game { _turn = Villagers }
+        game'   = game { _stage = VillagersTurn }
         n       = length $ game' ^. players
 
 prop_checkVillagersTurnLynchesNoOneWhenConflictedAndNoScapegoats :: Game -> Property
 prop_checkVillagersTurnLynchesNoOneWhenConflictedAndNoScapegoats game =
     forAll (runArbitraryCommands n game') $ \game'' ->
     length (last $ groupSortOn (length . flip elemIndices (Map.elems $ game'' ^. votes)) (nub . Map.elems $ game'' ^. votes)) > 1
-    ==> length (filterDead $ run_ checkTurn game'' ^. players) == length (filterDead $ game' ^. players)
+    ==> length (filterDead $ run_ checkStage game'' ^. players) == length (filterDead $ game' ^. players)
     where
-        game'   = (foldl killPlayer game (filterScapegoats $ game ^. players)) { _turn = Villagers }
+        game'   = (foldl killPlayer game (filterScapegoats $ game ^. players)) { _stage = VillagersTurn }
         n       = length $ game' ^. players
 
 prop_checkVillagersTurnLynchesScapegoatWhenConflicted :: Game -> Property
@@ -126,71 +126,71 @@ prop_checkVillagersTurnLynchesScapegoatWhenConflicted game =
     forAll (runArbitraryCommands n game') $ \game'' -> and [
         length (last $ groupSortOn (length . flip elemIndices (Map.elems $ game'' ^. votes)) (nub . Map.elems $ game'' ^. votes)) > 1,
         any isScapegoat $ game' ^. players
-        ] ==> isScapegoat $ head (filterDead $ run_ checkTurn game'' ^. players)
+        ] ==> isScapegoat $ head (filterDead $ run_ checkStage game'' ^. players)
     where
-        game'   = game { _turn = Villagers }
+        game'   = game { _stage = VillagersTurn }
         n       = length $ game' ^. players
 
 prop_checkVillagersTurnResetsVotes :: Game -> Property
 prop_checkVillagersTurnResetsVotes game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    Map.null $ run_ checkTurn game'' ^. votes
+    Map.null $ run_ checkStage game'' ^. votes
     where
-        game'   = game { _turn = Villagers }
+        game'   = game { _stage = VillagersTurn }
         n       = length $ game' ^. players
 
 prop_checkVillagersTurnDoesNothingUnlessAllVoted :: Game -> Property
 prop_checkVillagersTurnDoesNothingUnlessAllVoted game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isVillagersTurn $ run_ checkTurn game''
+    isVillagersTurn $ run_ checkStage game''
     where
-        game'   = game { _turn = Villagers }
+        game'   = game { _stage = VillagersTurn }
         n       = length (game' ^. players) - 1
 
 prop_checkWerewolvesTurnAdvancesToVillagers :: Game -> Property
 prop_checkWerewolvesTurnAdvancesToVillagers game =
     forAll (runArbitraryCommands n game') $ \game' ->
-    isVillagersTurn $ run_ checkTurn game'
+    isVillagersTurn $ run_ checkStage game'
     where
-        game'   = game { _turn = Werewolves }
+        game'   = game { _stage = WerewolvesTurn }
         n       = length . filterWerewolves $ game' ^. players
 
 prop_checkWerewolvesTurnKillsOnePlayerWhenConsensus :: Game -> Property
 prop_checkWerewolvesTurnKillsOnePlayerWhenConsensus game =
     forAll (runArbitraryCommands n game') $ \game'' ->
     length (last $ groupSortOn (length . flip elemIndices (Map.elems $ game'' ^. votes)) (nub . Map.elems $ game'' ^. votes)) == 1
-    ==> length (filterDead $ run_ checkTurn game'' ^. players) == 1
+    ==> length (filterDead $ run_ checkStage game'' ^. players) == 1
     where
-        game'   = game { _turn = Werewolves }
+        game'   = game { _stage = WerewolvesTurn }
         n       = length $ game' ^. players
 
 prop_checkWerewolvesTurnKillsNoOneWhenConflicted :: Game -> Property
 prop_checkWerewolvesTurnKillsNoOneWhenConflicted game =
     forAll (runArbitraryCommands n game') $ \game'' ->
     length (last $ groupSortOn (length . flip elemIndices (Map.elems $ game'' ^. votes)) (nub . Map.elems $ game'' ^. votes)) > 1
-    ==> length (filterDead $ run_ checkTurn game'' ^. players) == 0
+    ==> length (filterDead $ run_ checkStage game'' ^. players) == 0
     where
-        game'   = game { _turn = Werewolves }
+        game'   = game { _stage = WerewolvesTurn }
         n       = length $ game' ^. players
 
 prop_checkWerewolvesTurnResetsVotes :: Game -> Property
 prop_checkWerewolvesTurnResetsVotes game =
     forAll (runArbitraryCommands n game') $ \game' ->
-    Map.null $ run_ checkTurn game' ^. votes
+    Map.null $ run_ checkStage game' ^. votes
     where
-        game'   = game { _turn = Werewolves }
+        game'   = game { _stage = WerewolvesTurn }
         n       = length . filterWerewolves $ game' ^. players
 
 prop_checkWerewolvesTurnDoesNothingUnlessAllVoted :: Game -> Property
 prop_checkWerewolvesTurnDoesNothingUnlessAllVoted game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isWerewolvesTurn $ run_ checkTurn game''
+    isWerewolvesTurn $ run_ checkStage game''
     where
-        game'   = game { _turn = Werewolves }
+        game'   = game { _stage = WerewolvesTurn }
         n       = length (filterWerewolves $ game' ^. players) - 1
 
-prop_checkGameOverAdvancesTurn :: Game -> Property
-prop_checkGameOverAdvancesTurn game =
+prop_checkGameOverAdvancesStage :: Game -> Property
+prop_checkGameOverAdvancesStage game =
     forAll (sublistOf $ game ^. players) $ \players' ->
     let game' = foldl killPlayer game players' in
         length (nub . map (_allegiance . _role) . filterAlive $ game' ^. players) <= 1
@@ -204,10 +204,10 @@ prop_checkGameOverDoesNothingWhenAtLeastTwoAllegiancesAlive game =
             length (nub . map (_allegiance . _role) . filterAlive $ game' ^. players) > 1
             ==> not . isGameOver $ run_ checkGameOver game'
 
-prop_startGameStartsWithNightfallTurn :: [Player] -> Property
-prop_startGameStartsWithNightfallTurn players =
+prop_startGameStartsWithSunsetStage :: [Player] -> Property
+prop_startGameStartsWithSunsetStage players =
     isRight (runExcept . runWriterT $ startGame "" players)
-    ==> isNightfallTurn (fst . fromRight . runExcept . runWriterT $ startGame "" players)
+    ==> isSunset (fst . fromRight . runExcept . runWriterT $ startGame "" players)
 
 prop_startGameUsesGivenPlayers :: [Player] -> Property
 prop_startGameUsesGivenPlayers players' = and [